requirements.txt
flask
pillow
google-generativeai
python-dotenv
flask-cors